pyproject.toml
[tool.poetry]
name = "drf-localize"
version = "0.0.1"
description = "Localization Experience"
authors = ["Dorin Musteața"]

[tool.poetry.dependencies]
python = ">=3.7 <=3.10"
djangorestframework = "^3.12.4"
Django = "^3.2.8"
drf-yasg = "^1.20.0"
jsonschema = "^3.2.0"
django-cors-headers = "^3.10.0"
django-filter = "^21.1"
swagger_spec_validator = "^2.7.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
